# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Apex Inventory & Order Management API
  version: 1.1.0
  description: |
    Week 3 bench project — products, inventory, and orders with atomic stock reservation,
    fulfillment, pagination, and idempotent creates.
  contact:
    name: alerodriguez162
servers:
  - url: http://localhost:3001
    description: Local development
paths:
  /api/health:
    get:
      tags: [Ops]
      summary: Health check
      responses:
        '200':
          description: Service is up
  /api/v1:
    get:
      tags: [Ops]
      summary: Resource map
      responses:
        '200':
          description: Available resources and conventions
  /api/v1/products:
    get:
      tags: [Products]
      summary: List products
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: q
          in: query
          schema: { type: string }
        - name: sku
          in: query
          schema: { type: string }
        - name: sort
          in: query
          schema:
            type: string
            enum:
              - created_at_desc
              - created_at_asc
              - name_asc
              - name_desc
              - price_asc
              - price_desc
      responses:
        '200':
          description: Paginated products
    post:
      tags: [Products]
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProduct'
      responses:
        '201':
          description: Created
        '400':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
  /api/v1/products/{id}:
    get:
      tags: [Products]
      summary: Get product by id
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Product
        '404':
          $ref: '#/components/responses/Error'
    patch:
      tags: [Products]
      summary: Update product fields
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProduct'
      responses:
        '200':
          description: Updated product
        '400':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
  /api/v1/inventory:
    get:
      tags: [Inventory]
      summary: List inventory
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: lowStock
          in: query
          schema: { type: boolean }
        - name: threshold
          in: query
          schema: { type: integer, minimum: 0 }
      responses:
        '200':
          description: Paginated inventory
  /api/v1/inventory/{sku}:
    get:
      tags: [Inventory]
      summary: Get stock by SKU
      parameters:
        - name: sku
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Inventory row
        '404':
          $ref: '#/components/responses/Error'
    patch:
      tags: [Inventory]
      summary: Adjust stock
      parameters:
        - name: sku
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateInventory'
      responses:
        '200':
          description: Updated inventory
        '400':
          $ref: '#/components/responses/Error'
  /api/v1/orders:
    get:
      tags: [Orders]
      summary: List orders
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, cancelled]
      responses:
        '200':
          description: Paginated orders
    post:
      tags: [Orders]
      summary: Create order (reserves stock)
      parameters:
        - name: Idempotency-Key
          in: header
          schema: { type: string, minLength: 8, maxLength: 128 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrder'
      responses:
        '201':
          description: Created
        '200':
          description: Idempotent replay
        '409':
          $ref: '#/components/responses/Error'
  /api/v1/orders/{id}:
    get:
      tags: [Orders]
      summary: Get order by id
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Order
        '404':
          $ref: '#/components/responses/Error'
  /api/v1/orders/{id}/cancellation:
    post:
      tags: [Orders]
      summary: Cancel order and release reserved stock
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Cancelled order
  /api/v1/orders/{id}/fulfillment:
    post:
      tags: [Orders]
      summary: Fulfill order (deduct on-hand + reserved stock)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Fulfilled order
        '409':
          $ref: '#/components/responses/Error'
  /api/v1/stats:
    get:
      tags: [Ops]
      summary: Aggregate inventory and order stats
      parameters:
        - name: threshold
          in: query
          schema: { type: integer, minimum: 0, default: 5 }
      responses:
        '200':
          description: Stats snapshot
components:
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
  schemas:
    CreateProduct:
      type: object
      required: [sku, name, priceCents]
      properties:
        sku: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        priceCents: { type: integer, minimum: 0 }
        initialStock: { type: integer, minimum: 0 }
    UpdateProduct:
      type: object
      properties:
        name: { type: string }
        description: { type: string, nullable: true }
        priceCents: { type: integer, minimum: 0 }
    UpdateInventory:
      type: object
      properties:
        quantity: { type: integer, minimum: 0 }
        delta: { type: integer }
    CreateOrder:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          items:
            type: object
            required: [sku, quantity]
            properties:
              sku: { type: string }
              quantity: { type: integer, minimum: 1 }
    ErrorBody:
      type: object
      required: [error, code, requestId]
      properties:
        error: { type: string }
        code: { type: string }
        requestId: { type: string }
        details: {}
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'